+++ scripts/internal/winmake.py
@@ -29,15 +29,18 @@ import tempfile
 
 APPVEYOR = bool(os.environ.get('APPVEYOR'))
 PYTHON = sys.executable if APPVEYOR else os.getenv('PYTHON', sys.executable)
-RUNNER_PY = 'pyftpdlib\\test\\runner.py'
 GET_PIP_URL = "https://bootstrap.pypa.io/get-pip.py"
 PY3 = sys.version_info[0] >= 3
+PYTEST_ARGS = "-v --tb=native "
+if PY3:
+    PYTEST_ARGS += "-o "
 HERE = os.path.abspath(os.path.dirname(__file__))
 ROOT_DIR = os.path.realpath(os.path.join(HERE, "..", ".."))
 PYPY = '__pypy__' in sys.builtin_module_names
 DEPS = [
     "pip",
     "psutil",
+    "pytest",
     "pyopenssl",
     "pypiwin32",
     "setuptools",
@@ -369,10 +372,10 @@ def lint():
     sh("%s -m flake8 %s" % (PYTHON, py_files), nolog=True)
 
 
-def test(name=RUNNER_PY):
+def test(args=""):
     """Run tests."""
     build()
-    sh("%s %s" % (PYTHON, name))
+    sh("%s -m pytest %s %s" % (PYTHON, PYTEST_ARGS, args))
 
 
 def test_authorizers():
@@ -413,7 +416,7 @@ def test_servers():
 def coverage():
     """Run coverage tests."""
     build()
-    sh("%s -m coverage run %s" % (PYTHON, RUNNER_PY))
+    sh("%s -m coverage run -m pytest %s" % (PYTHON, PYTEST_ARGS))
     sh("%s -m coverage report" % PYTHON)
     sh("%s -m coverage html" % PYTHON)
     sh("%s -m webbrowser -t htmlcov/index.html" % PYTHON)
@@ -422,13 +425,13 @@ def coverage():
 def test_by_name(name):
     """Run test by name."""
     build()
-    sh("%s -m unittest -v %s" % (PYTHON, name))
+    test(name)
 
 
 def test_failed():
     """Re-run tests which failed on last run."""
     build()
-    sh("%s %s --last-failed" % (PYTHON, RUNNER_PY))
+    sh("%s -m pytest %s --last-failed" % (PYTHON, PYTEST_ARGS))
 
 
 def install_git_hooks():